+++ src/browser/index.ts
@@ -1,4 +1,5 @@
 /// <amd-module name='graphqlade/dist/browser'/>
+export * from "../client/GraphQLClient";
 export * from "../util/AsyncPushIterator";
 export * from "../ws/GraphQLClientWebSocket";
 export * from "../ws/GraphQLWebSocketClient";