+++ src/client/index.ts
@@ -1,0 +1,2 @@
+export * from "./GraphQLClient";
+export * from "./GraphQLRequestError";